+++ Inshort Clone/public/favicon.ico
