+++ Inshort Clone/public/manifest.json
