+++ Inshort Clone/public/robots.txt
