+++ Inshort Clone/src/App.js
