+++ Inshort Clone/src/components/Footer/Footer.css
@@ -1,0 +1,17 @@
+.footer {
+    background-color: #222;
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+    color: white;
+  }
+  .name {
+    font-size: 20px;
+    padding: 20px 30px 15px;
+    color: white;
+    text-decoration: none;
+  }
+  
+  a {
+    color: white;
+  }
