+++ Inshort Clone/src/components/NewsCard/NewsCard.css
@@ -1,0 +1,83 @@
+@import url("https://fonts.googleapis.com/css2?family=Roboto:wght@300&display=swap");
+
+.newsCard{
+    display: flex;
+    border-radius: 5px;
+    margin-bottom: 25px;
+    padding: 10px;
+    box-shadow: 0 2px 5px 0 rgba(0, 0, 0, 0.16), 0 2px 10px 0 rgba(0, 0, 0, 0.12);
+    width: 100%;
+}
+
+.newsImage{
+    width: 320px;
+    height: 270px;
+    object-fit: cover;
+}
+
+.newsText{
+    display: flex;
+    flex-direction: column;
+    justify-content: space-between;
+    padding: 0 0 0 10px;
+}
+
+.title{
+    font-size: 22px;
+    line-height: 27px;
+    font-family: 'Roboto',sans-serif;
+}
+
+.author{
+    font-size: 12px;
+    color: #808290;
+    line-height: 22px;
+    font-weight: 300;
+}
+
+.short{
+    font-family: 'Conv_ClanPro-Bold',Sans-Serif;
+    color: #44444d;
+}
+
+.muted{
+    font-weight: 100;
+}
+
+.lowerNewsText{
+    display: flex;
+    flex-direction: column;
+    justify-content: space-between;
+    flex: 1;
+    padding: 10px 0;
+}
+
+.description{
+    font-size: 16px;
+    line-height: 22px;
+    color: #44444d;
+}
+
+.readmore{
+    font-size: 12px;
+    padding-top: 10px;
+    font-weight: 400;
+}
+
+
+.source{
+    color: black;
+    text-decoration: none;
+}
+
+@media (max-width:800px) {
+    .newsImage{
+        width: 100%;
+        height: 100%;
+        padding-bottom: 20px;
+    }
+    .newsCard{
+        flex-direction: column;
+        width: 100%;
+    }
+}
