+++ Inshort Clone/src/data/category.js
@@ -1,0 +1,11 @@
+const categories = [
+    "business",
+    "entertainment",
+    "general",
+    "health",
+    "science",
+    "sports",
+    "technology"
+];
+
+export default categories;
