--- src/components/Footer/Footer.css
@@ -1,17 +1,0 @@
-.footer {
-    background-color: #222;
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-    color: white;
-  }
-  .name {
-    font-size: 20px;
-    padding: 20px 30px 15px;
-    color: white;
-    text-decoration: none;
-  }
-  
-  a {
-    color: white;
-  }
