--- src/components/Navinshorts.css
@@ -1,10 +1,0 @@
-.nav{
-    height: 71px;
-    box-shadow: 0px 1px 10px rgba(0,0,0,.3);
-    display: flex;
-    align-items: center;
-}
-
-.icon{
-    flex: 0.48;
-}
