--- src/components/NewsContent/NewsContent.css
@@ -1,40 +1,0 @@
-.content{
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-    padding: 20px;
-}
-
-.downloadMessage{
-    width: 94%;
-    display: flex;
-    align-items: center;
-    background: #f44336;
-    height: 48px;
-    padding:  0 40px;
-    margin: 20px 0 35px 0 ;
-    color: white;
-    box-shadow: 0 2px 5px 0 rgba(0,0,0,0.16), 0 2px 10px 0 rgba(0, 0, 0, 0.12) ;
-}
-
-.downloadText{
-    flex: 1;
-}
-
-@media (max-width: 995px){
-    .downloadMessage{
-        display: none;
-    }
-}
-
-.btn{
-    border-radius: 5px;
-    box-shadow: 0 2px 5px 0 rgba(0, 0, 0, 0.16), 0 2px 10px 0 rgba(0, 0, 0, 0.12);
-    font-size: 15px;
-    color: black;
-    cursor: pointer;
-    border: none;
-    background-color: white;
-    padding: 10px 30px;
-    margin: 12px 0 20px 0;
-}
