--- src/data/category.js
@@ -1,11 +1,0 @@
-const categories = [
-    "business",
-    "entertainment",
-    "general",
-    "health",
-    "science",
-    "sports",
-    "technology"
-];
-
-export default categories;
